feat: move swapQutote to upper directory
