+++ packages/swapper/src/swappers/zrx/ZrxSwapper.test.ts
@@ -5,10 +5,11 @@ import Web3 from 'web3'
 
 import { ZrxSwapper } from '..'
 import { FOX } from '../utils/test-data/assets'
+import { setupQuote } from '../utils/test-data/setupSwapQuote'
 import { zrxBuildTrade } from '../zrx/zrxBuildTrade/zrxBuildTrade'
 import { getZrxTradeQuote } from './getZrxTradeQuote/getZrxTradeQuote'
 import { getUsdRate } from './utils/helpers/helpers'
-import { setupBuildTrade, setupExecuteTrade, setupQuote } from './utils/test-data/setupSwapQuote'
+import { setupBuildTrade, setupExecuteTrade } from './utils/test-data/setupZrxSwapQuote'
 import { ZrxApprovalNeeded } from './ZrxApprovalNeeded/ZrxApprovalNeeded'
 import { ZrxApproveInfinite } from './ZrxApproveInfinite/ZrxApproveInfinite'
 import { zrxExecuteTrade } from './zrxExecuteTrade/zrxExecuteTrade'
