+++ packages/swapper/src/swappers/zrx/utils/test-data/setupZrxSwapQuote.ts
@@ -1,35 +1,7 @@
 import { Asset } from '@shapeshiftoss/types'
 
-import { GetTradeQuoteInput, TradeQuote } from '../../../../api'
 import { FOX, WETH } from '../../../utils/test-data/assets'
 import { ZrxQuoteResponse } from '../../types'
-
-export const setupQuote = () => {
-  const sellAsset: Asset = { ...FOX }
-  const buyAsset: Asset = { ...WETH }
-  const tradeQuote: TradeQuote<'eip155:1'> = {
-    buyAmount: '',
-    sellAmount: '1000000000000000000',
-    sellAsset,
-    buyAsset,
-    allowanceContract: 'allowanceContractAddress',
-    sellAssetAccountId: '0',
-    minimum: '0',
-    maximum: '999999999999',
-    feeData: { fee: '0', tradeFee: '0', chainSpecific: {} },
-    rate: '1',
-    sources: []
-  }
-
-  const quoteInput: GetTradeQuoteInput = {
-    sellAmount: '1000000000000000000',
-    sellAsset,
-    buyAsset,
-    sellAssetAccountId: '0',
-    sendMax: false
-  }
-  return { quoteInput, tradeQuote, buyAsset, sellAsset }
-}
 
 export const setupZrxTradeQuoteResponse = () => {
   const sellAsset: Asset = { ...FOX }
